Code correction

+++ CoditechLicenseApplication.DataAccessLayer/DataEntity/ApplicationLicenseDetail.cs
@@ -21,9 +21,9 @@ namespace Coditech.DataAccessLayer.DataEntity
         public System.DateTime ValidFromDate { get; set; }
         public System.DateTime ValidUptoDate { get; set; }
         public bool IsActive { get; set; }
-        public Nullable<long> CreatedBy { get; set; }
+        public Nullable<int> CreatedBy { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
-        public Nullable<long> ModifiedBy { get; set; }
+        public Nullable<int> ModifiedBy { get; set; }
         public Nullable<System.DateTime> ModifiedDate { get; set; }
     }
 }
